requirements.txt
Faker==18.3.1
kafka-python==2.0.2
python-dateutil==2.8.2
six==1.16.0
